requirements.txt
streamlit
groq
langchain
langchain-groq
python-dotenv
beautifulsoup4
faiss-cpu
langchain-google-genai
langchain-community